introduce `async_mining_allowed` as JDS config parameter
before issue #853 was reported, JDS would simply support async jobs by default and completely ignore this flag.

but checking this flag implies that JDS could either support async jobs or not, and that is what this commit does.

a new `asyn_mining_allowed` parameter is introduced to the TOML config files, and that is used when:
- checking for the flags of `SetupConnection` messages
- responding to `AllocateMiningJobToken` messages